requirements.txt
Django~=5.0.6
djangorestframework~=3.15.1
channels~=4.1.0
pillow~=10.3.0
asgiref~=3.8.1
channels-redis